+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+- Renamed Changelog to Release Notes in Settings.
+- Updated icon for tags to be more clear.
+
 ## [1.1.0] - 2018-07-16
 
 ### Added